extract out projects for cleaner code
